requirements.txt
Django==4.1
django-ckeditor==6.4.0
django_import_export==2.8.0
djangorestframework==3.13.1
drf_yasg==1.20.0
pycryptodome==3.15.0
requests==2.28.1
Pillow==9.1.0
